Find a way to display user costs

+++ src/main/java/com/costshare/repository/UserCostRepository.java
@@ -15,5 +15,5 @@ import java.util.List;
 @Repository
 public interface UserCostRepository extends JpaRepository<UserCost, Long> {
 
-    public List<UserCost> findAllByBaseCost(Long id);
+    public List<UserCost> findAllByBaseCost(int id);
 }
